+++ components/Chat/Chat.tsx
@@ -1,27 +1,10 @@
 import ChatInput from "./ChatInput";
 import Messages from "./Messages";
-import { useAccount } from "wagmi";
-import useConnectWallet from "@/hooks/useConnectWallet";
 import { useChatProvider } from "@/providers/ChatProvider";
 
 const Chat = () => {
-  const { connectWallet } = useConnectWallet();
-  const { address } = useAccount();
-  const {
-    messages,
-    input,
-    handleInputChange,
-    handleSubmit: handleChatSubmit,
-  } = useChatProvider();
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    if (!address) {
-      e.preventDefault();
-      await connectWallet();
-      return;
-    }
-    handleChatSubmit(e);
-  };
+  const { messages, input, handleInputChange, handleSubmit } =
+    useChatProvider();
 
   return (
     <div className="w-full items-center flex flex-col">
